Update config for the Style/AndOr cop
"and" and "or" have the same logical meaning as "&&" and "||", but with
a lower precedence. Outside of the context of comparisons, this is a
language feature that can actually be quite useful, e.g. when returning
early from a Rails controller.
I'm therefore suggesting to enforce replacing "and" and "or" only in the
context of conditionals, rather then always.