[#19] Exclude config ignored? (#21)
* change to merging defaults

Since rubucoop 0.57.0 one does no longer have to fully override the
AllCops:Include and AllCops:Exclude lists, but can instead merge them.

* Update the version of rubocop we require

Since 0.57.0, it's possible to "merge" the default configuration for
e.g. AllCops:Include - which is what we usually like (otherwise our gem
would need to fully override the whole list...)

* Update list of excluded files

Since we're now requiring a rubocop version starting fromm 0.57.0, we
can drop the things that are in the default configuration (see here:
https://github.com/rubocop-hq/rubocop/blob/5d5e7c0f8e195315d10b165e11399b6641754ce2/config/default.yml).

Instead we can now make use of "inherit_mode" to merge the configuration
with the defaults.

* Add a post install message

I added a post install message to remind users we're changing the
configuration for which directories to include/exclude.

* Remove redundant configuration

This removes the `DisplayCopNames` rule. It defaults to true anyways, so
there's not need to define it on our side as well.

closes #19